Deleted LoginPage.js
Deleted LoginPage.js and rename LoginPageAlter.js
